server/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pandas
beautifulsoup4
playwright
xmltodict
pydantic
